Engine: First rendering on material factory support and shader refactoring.

+++ src/Alimer/Graphics/D3D12/D3D12RenderPipeline.cs
@@ -66,7 +66,7 @@ internal unsafe class D3D12RenderPipeline : RenderPipeline
         };
         for (uint binding = 0; binding < descriptor.VertexBufferLayouts.Length; binding++)
         {
-            ref readonly VertexBufferLayout layout = ref descriptor.VertexBufferLayouts[binding];
+            ref readonly VertexBufferLayout layout = ref descriptor.VertexBufferLayouts[(int)binding];
 
             if (layout.Stride == 0)
                 continue;
